requirements.txt
# AI Chatbot System Requirements
# Frontend
streamlit>=1.44.0
streamlit-chat>=0.1.1

# Backend
Flask>=3.1.0
Flask-RESTful>=0.3.10
Flask-Cors>=5.0.0
Flask-JWT-Extended>=4.7.0
bcrypt>=4.3.0
python-dotenv>=1.1.0
gunicorn>=23.0.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=1.4.0,<1.5.0
alembic>=1.15.2

# NLP
rasa>=3.6.21
spacy>=3.8.5

# Utilities
requests>=2.32.3
loguru>=0.7.3
pydantic<1.10.10
pytest>=8.3.5

# Security
cryptography>=44.0.2